Add GM mode
:D

+++ src/mail.c
@@ -390,9 +390,13 @@ struct mail_info *info;
     message_seen = TRUE;
     verbalize("%s, %s!  %s.", Hello(md), plname, info->display_txt);
 
+
     if (info->message_typ) {
 	struct obj *obj = mksobj(SCR_MAIL, FALSE, FALSE);
 	if (!obj) return;
+
+	/* In game master mode, mail from me will spawn monsters. Read it immediately, so you can't avoid them :P --Amy */
+	if (flags.gmmode) readmail(obj);
 	if (distu(md->mx,md->my) > 2)
 	    verbalize("Catch!");
 	display_nhwindow(WIN_MESSAGE, FALSE);
@@ -543,14 +547,27 @@ struct obj *otmp;
 		  
 		  if (!msg)
 		    goto bail;
-		  
+
 		  *msg = '\0';
 		  msg++;
-		  
+
 		  pline ("This message is from '%s'.", curline);
+
+		  if (flags.gmmode && !(flags.gmmessage)) goto dontread;
 
 		  msg[strlen(msg) - 1] = '\0'; /* kill newline */
 		  pline ("It reads: \"%s\".", msg);
+
+dontread:
+
+		/* Game master mode: if active, I (Amy) can send you mail to spawn monsters!
+		 * If you turned off the gmmessage option, the monster I selected will be a surprise :) */
+		  if (flags.gmmode && strlen(curline) == 7 && !strncmpi(curline, "AmyBSOD", 7)) {
+			if (strlen(msg) > 9 && !strncmpi(msg, "genesis ", 8)) {
+				u.gmmailsreceived++;
+				gmmode_genesis(msg + 8);
+			}
+		  }
 
 		  seen_one_already = TRUE;
 		  errno = 0;
